Fix ResNet build.
