+++ CIFAR/ResNet.swift
@@ -117,42 +117,3 @@ extension ResNet {
         self.init(blockCount: kind.rawValue)
     }
 }
-
-// TODO: remove this when TF supports differentiableReduce, thanks @rxwei!
-public extension Array where Element: Differentiable {
-    func differentiableReduce<Result: Differentiable>(
-        _ initialResult: Result,
-        _ nextPartialResult: @differentiable (Result, Element) -> Result
-    ) -> Result {
-        return reduce(initialResult, nextPartialResult)
-    }
-
-    @usableFromInline
-    @differentiating(differentiableReduce(_:_:), wrt: (self, initialResult))
-    internal func reduceDerivative<Result: Differentiable>(
-        _ initialResult: Result,
-        _ nextPartialResult: @differentiable (Result, Element) -> Result
-    ) -> (value: Result,
-          pullback: (Result.CotangentVector) -> (Array.CotangentVector, Result.CotangentVector)) {
-        var pullbacks: [(Result.CotangentVector) -> (Result.CotangentVector, Element.CotangentVector)] = []
-        let count = self.count
-        pullbacks.reserveCapacity(count)
-        var result = initialResult
-        for element in self {
-            let (y, pb) = Swift.valueWithPullback(at: result, element, in: nextPartialResult)
-            result = y
-            pullbacks.append(pb)
-        }
-        return (value: result, pullback: { cotangent in
-            var resultCotangent = cotangent
-            var elementCotangents = CotangentVector([])
-            elementCotangents.base.reserveCapacity(count)
-            for pullback in pullbacks.reversed() {
-                let (newResultCotangent, elementCotangent) = pullback(resultCotangent)
-                resultCotangent = newResultCotangent
-                elementCotangents.base.append(elementCotangent)
-            }
-            return (CotangentVector(elementCotangents.base.reversed()), resultCotangent)
-        })
-    }
-}
